fix: made searchbar ssr
