+++ .astro/settings.json
@@ -1,6 +1,6 @@
 {
   "_variables": {
-    "lastUpdateCheck": 1725876809059
+    "lastUpdateCheck": 1727019530972
   },
   "devToolbar": {
     "enabled": false
