+++ src/pages/app/pandals.astro
@@ -4,7 +4,7 @@ import { MdOutlineFestival } from 'react-icons/md';
 import Layout from '@/layouts/WebAppLayout.astro';
 import { cn } from '@/libs/utils';
 import AppHeader from '@/components/app/shared/AppHeader.astro';
-import { SearchBar } from '@/components/app/shared/SearchBar';
+import SearchBar from '@/components/app/shared/Search/SearchBar.astro';
 import { PANDAL_MARKER_ICON } from '@/constants/location';
 import { PANDALS } from '@/constants/pandals';
 
@@ -26,7 +26,7 @@ const pandals = PANDALS.map((pandal, index) => ({
     )}
   >
     <AppHeader />
-    <SearchBar client:load client:only="react" />
+    <SearchBar />
 
     <div
       class="flex flex-col justify-center px-2 py-2 bg-[#e6dfcf] rounded-[22px] drop-shadow-lg shadow-black"
