+++ src/stores/search/index.ts
@@ -1,0 +1,3 @@
+import { atom } from 'nanostores';
+
+export const searchStore = atom<string>('');
